+++ lib/super/filter/plugin.rb
@@ -13,11 +13,15 @@ module Super
   class Controls
     module Optional
       def filters_enabled?
-        actual.respond_to?(:filter_schema)
+        true
       end
 
       def filter_schema
-        actual.filter_schema
+        default_for(:filter_schema) do
+          Filter.new do |fields, type|
+            Filter::Guesser.new(model: model, fields: fields, type: type).call
+          end
+        end
       end
     end